split the mega file into classes
