+++ BuyRecipes/Domain/BaseRecipeAquisition.cs
@@ -1,0 +1,13 @@
+﻿namespace Denifia.Stardew.BuyRecipes.Domain
+{
+    public abstract class BaseRecipeAquisition
+    {
+        public BaseRecipeAquisition()
+        {
+        }
+
+        public BaseRecipeAquisition(string data)
+        {
+        }
+    }
+}
