+++ BuyRecipes/Domain/GameItem.cs
@@ -1,0 +1,8 @@
+﻿namespace Denifia.Stardew.BuyRecipes.Domain
+{
+    public class GameItem
+    {
+        public string Name { get; set; }
+        public int Id { get; set; }
+    }
+}
