+++ BuyRecipes/Domain/GameItemWithQuantity.cs
@@ -1,0 +1,7 @@
+﻿namespace Denifia.Stardew.BuyRecipes.Domain
+{
+    public class GameItemWithQuantity : GameItem
+    {
+        public int Quantity { get; set; }
+    }
+}
