+++ BuyRecipes/Domain/IRecipeAquisitionConditions.cs
@@ -1,0 +1,8 @@
+﻿namespace Denifia.Stardew.BuyRecipes.Domain
+{
+    public interface IRecipeAquisitionConditions
+    {
+        bool AcceptsConditions(string condition);
+        int Cost { get; }
+    }
+}
